requirements.txt
openai>=1.0.0,<2.0.0
PyYAML>=6.0,<7.0
